impl asio and pthread instead of sfml-net and sys, intuitive naming
