+++ components/impl/impl.h
@@ -6,11 +6,29 @@
 #define IMPL_H
 
 #ifdef ESP_PLATFORM
-#include "esp_impl.h"
+
+#define logi(tag, fmt, ...) ESP_LOGI(tag, fmt, ##__VA_ARGS__)
+#ifdef LOG_DEBUG
+#define logd(tag, fmt, ...) ESP_LOGD(tag, fmt, ##__VA_ARGS__)
+#else
+#define logd(tag, fmt, ...)
+#endif
+#define loge(tag, fmt, ...) ESP_LOGE(tag, fmt, ##__VA_ARGS__)
+#include "freertos_impl.h"
 
 #else
-#include "sfml_impl.h"
+
+#define logi(tag, fmt, ...) printf((std::string("%s: ") + fmt + std::string("\n")).c_str(), tag, ##__VA_ARGS__)
+#ifdef LOG_DEBUG
+#define logd(tag, fmt, ...) printf((std::string("%s: ") + fmt + std::string("\n")).c_str(), tag, ##__VA_ARGS__)
+#else
+#define logd(tag, fmt, ...)
+#endif
+#define loge(tag, fmt, ...) fprintf(stderr, (std::string("%s: ") + fmt + std::string("\n")).c_str(), tag, ##__VA_ARGS__)
+#include "pthread_impl.h"
 
 #endif
+
+#include "asio_impl.h"
 
 #endif //IMPL_H
